fix base64 build issue
