+++ packages/local-mail-watcher/src/lib/base64.ts
@@ -1,0 +1,19 @@
+/**
+ * Utility functions for handling base64 encoding/decoding
+ */
+
+// Import types from our declaration file
+import '../types/node-buffer';
+
+/**
+ * Decodes a base64 string to UTF-8 text
+ * Uses Node.js Buffer but with type safety
+ */
+export function decodeBase64(base64String: string): string {
+  try {
+    return Buffer.from(base64String, 'base64').toString('utf-8');
+  } catch (error) {
+    console.error('Error decoding base64 string:', error);
+    return base64String; // Return original string if decoding fails
+  }
+}
